entries data changes

+++ srcs/ServerApi/Entries/API.Get.cs
@@ -11,7 +11,7 @@ namespace FriendlyCashFlow.API.Entries
          try
          {
             data.Sorting = Convert.ToInt64(data.SearchDate.Subtract(new DateTime(1901, 1, 1)).TotalDays);
-            data.Sorting += (data.EntryID / (decimal)Math.Pow(10, data.EntryID.ToString().Length));
+            data.Sorting += ((decimal)data.EntryID / (decimal)Math.Pow(10, data.EntryID.ToString().Length));
          }
          catch (Exception) { throw; }
       }
